(MAINT) Fix icon display
Prior to this commit the icons display at full width as
blocks - this made the text around them unreadable with
the disruption.

This commit modifies their definition and css to pin them
to 1em height and width until such a time as I have some
bandwidth to update the code and figure out what is wrong.